Rework internals
Major rework of the internal datastructure so that it performs better
with very large crates. Many associated bugfixes.

commit 17adf40589576ab8b5f18793450f779c11368a83
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat May 6 15:34:14 2023 +1000

    bugfix - pass in srcEntityId

commit 0ab0db332b0405368187a80748724018cd8ffee8
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat May 6 15:33:53 2023 +1000

    don't assume root dataset - require the src to be passed in

commit 111913b433faf75c7c6c7f4db9da8659f2f70acd
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat May 6 15:33:28 2023 +1000

    remove label prop

commit e6b4d7b7ede8ed5061e60305cf40e3aafdec06b2
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat May 6 13:12:48 2023 +1000

    bugfix - ensure @id url encoded

commit c6d37fa8460a4bc4f61830c19090efed4a2fb786
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat May 6 13:01:00 2023 +1000

    tweaking loading indicator

commit 623fb4d4c9cefdd3eaf47ab7b9013c92f414b9d7
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat May 6 12:07:56 2023 +1000

    update tests

commit a613132d38f217bb3e586a4b31f71621ef7147b2
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat May 6 12:07:51 2023 +1000

    bugfix

commit 1cf570d4d8fae455f632effb2d03d76203bbcee6
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat May 6 11:41:09 2023 +1000

    bugfixing

commit 4d8edf8f1b9474a47dd5a4205f3aa006dc91c8f8
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat May 6 11:40:49 2023 +1000

    refactor

commit 90a2352a281aa4cf3846f1c7d7ca509c81b9c1d0
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat May 6 10:59:39 2023 +1000

    display loading indicator

commit 9f67d30fe2487f5b9fa513d3253196ad2ebf83f1
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat May 6 10:59:07 2023 +1000

    disable resolving additional entity types and bugfix

commit aac5699242b4aaca8926d9b6f027e5e00caafd6b
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat May 6 10:58:44 2023 +1000

    revise docs

commit 775d1aedb64723fc0e0f518b8e5b59d23c51d19a
Author: Marco La Rosa <m@lr.id.au>
Date:   Wed May 3 19:07:35 2023 +1000

    revise to suit reworked internals

commit b762e154a683bcd139a980be5897a2ce0973e5af
Author: Marco La Rosa <m@lr.id.au>
Date:   Wed May 3 19:07:07 2023 +1000

    revise paginate list down

commit b56b4b6edc9c6181d13f3cd87e0891e0b28e9fa4
Author: Marco La Rosa <m@lr.id.au>
Date:   Wed May 3 19:06:31 2023 +1000

    rework crate-manager internals

    Base on entity and property classes that are self managing

commit fefbfe8783af7ae16279ced21ad3962dd169b39f
Author: Marco La Rosa <m@lr.id.au>
Date:   Sun Apr 30 12:01:00 2023 +1000

    wip

commit 81f383ec7de83cbb4f26776f26ced543fab7d55f
Author: Marco La Rosa <m@lr.id.au>
Date:   Sun Apr 30 10:11:45 2023 +1000

    wip

commit 3bd1f669efa93f76495ee969e678cc60454d3a2b
Author: Marco La Rosa <m@lr.id.au>
Date:   Sun Apr 30 10:11:32 2023 +1000

    refactor

commit f0cfc672d85cf2f42b2e5203045bba8bb7ac5512
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat Apr 29 14:34:21 2023 +1000

    implementing entity property handling

commit c80ef5e395abf5fb96ff8a6e9e531e4a595c5333
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat Apr 29 13:56:28 2023 +1000

    implement entity class and tests

commit c58428bd0583f6698aebf1b2b5878289b20134d0
Author: Marco La Rosa <m@lr.id.au>
Date:   Sat Apr 29 11:43:35 2023 +1000

    implement properties class and tests

    Manages entity property operations

+++ src/crate-builder/RenderEntity/Shell.component.vue
@@ -275,7 +275,7 @@ const $emit = defineEmits([
 ]);
 
 watch([() => props.entity, () => props.profile], (n, o) => {
-    if (n.describoId !== o.describoId) {
+    if (n[0].describoId !== o[0].describoId) {
         data.extraProperties = [];
         data.entity = {};
         data.tabs = [];
@@ -322,7 +322,7 @@ function init() {
     }
     let properties = {};
     let propertyNames = [];
-    Object.keys(entity.properties).map((k) => (propertyNames[k.toLowerCase()] = k));
+    Object.keys(entity.properties).forEach((k) => (propertyNames[k.toLowerCase()] = k));
     Object.keys(entity.properties)
         .map((k) => k.toLowerCase())
         .sort()
@@ -420,7 +420,7 @@ function saveCrate() {
 function createProperty(patch) {
     console.debug("Render Entity component: emit(create:property)", patch);
     if (props.configuration.mode === "embedded") {
-        props.crateManager.addProperty({ ...patch });
+        props.crateManager.addProperty({ describoId: props.entity.describoId, ...patch });
     } else {
         $emit("add:property", { ...patch, entityId: props.entity.describoId });
     }
@@ -432,7 +432,7 @@ function createProperty(patch) {
 function saveProperty(patch) {
     console.debug("Render Entity component: emit(save:property)", patch);
     if (props.configuration.mode === "embedded") {
-        props.crateManager.updateProperty({ ...patch });
+        props.crateManager.updateProperty({ describoId: props.entity.describoId, ...patch });
     } else {
         $emit("save:property", patch);
     }
@@ -444,7 +444,10 @@ function saveProperty(patch) {
 function deleteProperty(data) {
     console.debug("Render Entity component: emit(delete:property)", data);
     if (props.configuration.mode === "embedded") {
-        props.crateManager.deleteProperty({ propertyId: data.propertyId });
+        props.crateManager.deleteProperty({
+            describoId: props.entity.describoId,
+            propertyId: data.propertyId,
+        });
     } else {
         $emit("delete:property", { propertyId: data.propertyId });
     }
@@ -459,7 +462,11 @@ function createEntity(data) {
     console.debug("Render Entity component: emit(create:entity)", data);
     if (props.configuration.mode === "embedded") {
         if (dataType === "datapack") {
-            props.crateManager.ingestAndLink({ property, json: data });
+            props.crateManager.ingestAndLink({
+                srcEntityId: props.entity.describoId,
+                property,
+                json: data,
+            });
         } else {
             let entity = props.crateManager.addEntity({ entity: data.json });
             props.crateManager.linkEntity({
@@ -490,6 +497,7 @@ function linkEntity(data) {
     console.debug("Render Entity component: emit(link:entity)", data);
     if (props.configuration.mode === "embedded") {
         props.crateManager.linkEntity({
+            srcEntityId: props.entity.describoId,
             property: data.property,
             tgtEntityId: data.json.describoId,
         });
@@ -500,6 +508,7 @@ function linkEntity(data) {
     saveCrate();
 }
 function deleteEntity(data) {
+    if (data.describoId === "RootDataset") return;
     console.debug("Render Entity component: emit(delete:entity)", data);
     if (props.configuration.mode === "embedded") {
         props.crateManager.deleteEntity(data);
